Successfully optimized file structures.

+++ feature_engineering.py
@@ -1,5 +1,6 @@
+import os
 import pickle
-from data_loader import spectral_feature_loader
+from vad.data_loader import spectral_feature_loader
 
 train_set_path = './wavs/train'
 dev_set_path = './wavs/dev'
@@ -10,19 +11,22 @@ n_mfcc_list = [8, 12, 15, 20]
 
 # all deltas
 for n_mfcc in n_mfcc_list:
-    X_train, sample_lengths, Y_train = spectral_feature_loader(
-        train_set_path, train_label_path,
-        n_mfcc=n_mfcc,
-        use_first_order=True, use_third_order=True)
+    if not os.path.exists(''.join(['train_123_', str(n_mfcc), '.pkl'])):
+        X_train, sample_lengths, Y_train = spectral_feature_loader(
+            train_set_path, train_label_path,
+            n_mfcc=n_mfcc,
+            use_first_order=True, use_third_order=True)
 
-    X_dev, sample_lengths_dev, Y_dev = spectral_feature_loader(
-        dev_set_path, dev_label_path,
-        n_mfcc=n_mfcc,
-        use_first_order=True, use_third_order=True)
+        pickle.dump(
+            [X_train, sample_lengths, Y_train],
+            open(''.join(['train_123_', str(n_mfcc), '.pkl']), 'wb'))
 
-    pickle.dump(
-        [X_train, sample_lengths, Y_train],
-        open(''.join(['train_123_', str(n_mfcc), '.pkl']), 'wb'))
-    pickle.dump(
-        [X_dev, sample_lengths_dev, Y_dev],
-        open(''.join(['dev_123_', str(n_mfcc), '.pkl'], 'wb')))
+    if not os.path.exists(''.join(['dev_123_', str(n_mfcc), '.pkl'])):
+        X_dev, sample_lengths_dev, Y_dev = spectral_feature_loader(
+            dev_set_path, dev_label_path,
+            n_mfcc=n_mfcc,
+            use_first_order=True, use_third_order=True)
+
+        pickle.dump(
+            [X_dev, sample_lengths_dev, Y_dev],
+            open(''.join(['dev_123_', str(n_mfcc), '.pkl']), 'wb'))
